Addressed reviewer comments.
